Split User components up
